+++ chapter-1/chapter-1.rb
@@ -1,0 +1,26 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 1"
+puts "# File: calc.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts "Calculations using floats:"
+puts 1+2
+puts 1.0 + 2.0
+puts 2.0 * 3.0
+puts 5.0 - 8.0
+puts 9.0 / 2.0
+puts
+
+puts "Calculations using integers:"
+puts 1+2
+puts 2*3
+puts 5-8
+puts 9/2
+puts
+
+puts "Calculations using more complex syntax:"
+puts 5 * (12-8) + -15
+puts 98 + (59872 / (13*8)) * -52
+puts