+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - LICENSE
 - README.md
 - big-ruby-program.rb
+- chapter-1.rb [examples from chapter-1]